Added React.
Installed React with: npm i react react-dom
Installed remount with: npm i remount - this lets me render the ask component on a HTML page by adding <ask-component> to the page
Rails generated a HomeController and gave it a #show action and view.
Wrote a basic jsx component named Ask.jsx and added it to the #show page.
Defined the root patg as home#show