Added a front end that displays an answer. (Question has not been added yet)
Rails scafolded a Question controller, model, etc
Updated the Ask component to post a question (JSON string) to the rails Question controller and set the response on the page.
Dealt with CSRF to pass the csrfToken along with the question.
Still ignoring what the question is and giving the exact same response every time.
Also not utilizing the Question model yet. To be used to store questions and answers as needed.