Script to import the PDF and create embeddings
Created the model that will hold the local embeddings generated from the book (BookDatum)

Created the file that goes through the book, page by page, and creates the model instances along with the embeddings.
Setup dotenv for the openapi key.
Waiting for the payment for the openai api access to go through.